Move cron and preening code to server, add support for quests in cron
