+++ common/script/cron.js
@@ -1,3 +1,4 @@
+// TODO what can be moved to /website/src?
 /*
   ------------------------------------------------------
   Cron and time / day functions
